Menu, Enemy Ground Check, BUG IS BACK
Added a menu, changed spawner enemies left alive check to not include corpses, added ground check for enemies, enemies move slower in air, fixed bug where mass was applied twice for force calculations in player and enemies. The bug of getting stuck to the wall when walking into it is back.